+++ src/main/java/szewek/mcflux/MCFlux.java
@@ -1,13 +1,15 @@
 package szewek.mcflux;
 
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
+import com.google.gson.JsonPrimitive;
 import net.minecraftforge.common.capabilities.CapabilityManager;
+import net.minecraftforge.fml.common.Loader;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.SidedProxy;
 import net.minecraftforge.fml.common.discovery.ASMDataTable;
-import net.minecraftforge.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLInterModComms;
-import net.minecraftforge.fml.common.event.FMLLoadCompleteEvent;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+import net.minecraftforge.fml.common.event.*;
+import net.minecraftforge.fml.common.versioning.ComparableVersion;
 import net.minecraftforge.oredict.RecipeSorter;
 import szewek.mcflux.api.ex.Battery;
 import szewek.mcflux.api.ex.EnergyNBTStorage;
@@ -20,13 +22,14 @@ import szewek.mcflux.fluxable.InjectFluxable;
 import szewek.mcflux.fluxable.PlayerEnergy;
 import szewek.mcflux.fluxable.WorldChunkEnergy;
 import szewek.mcflux.network.MCFluxNetwork;
-import szewek.mcflux.util.IInjectRegistry;
-import szewek.mcflux.util.InjectRegistry;
-import szewek.mcflux.util.MCFluxCreativeTab;
-import szewek.mcflux.util.NBTSerializableCapabilityStorage;
+import szewek.mcflux.util.*;
 import szewek.mcflux.util.recipe.BuiltShapedRecipe;
 import szewek.mcflux.wrapper.InjectWrappers;
 
+import java.io.File;
+import java.io.InputStreamReader;
+import java.net.HttpURLConnection;
+import java.net.URL;
 import java.util.Set;
 
 import static net.minecraftforge.common.MinecraftForge.EVENT_BUS;
@@ -34,6 +37,7 @@ import static net.minecraftforge.common.MinecraftForge.EVENT_BUS;
 @SuppressWarnings("unused")
 @Mod(modid = R.MF_NAME, name = R.MF_FULL_NAME, version = R.MF_VERSION, useMetadata = true, guiFactory = R.GUI_FACTORY, dependencies = R.MF_DEPENDENCIES)
 public class MCFlux {
+	private static File MC_DIR;
 	static String NEWER_VERSION = "";
 	static boolean UPDATE_CHECK_FINISHED = false;
 	static final MCFluxCreativeTab MCFLUX_TAB = new MCFluxCreativeTab();
@@ -43,17 +47,20 @@ public class MCFlux {
 	@Mod.EventHandler
 	public void preInit(FMLPreInitializationEvent e) {
 		L.prepare(e.getModLog());
+		MC_DIR = e.getModConfigurationDirectory().getParentFile();
 		MCFluxConfig.makeConfig(e.getSuggestedConfigurationFile());
 		if (R.MF_VERSION.charAt(0) == '@')
 			L.warn("You are running Minecraft-Flux with an unknown version (development maybe?)");
 		else
 			L.info("Minecraft-Flux " + R.MF_VERSION);
+		if (MCFluxConfig.UPDATE_CHECK)
+			new Thread(MCFlux::updateCheck, "MCFlux Update Check").start();
 		MCFluxNetwork.registerAll();
 		CapabilityManager cm = CapabilityManager.INSTANCE;
 		cm.register(IEnergy.class, new EnergyNBTStorage(), Battery::new);
 		cm.register(IFlavorEnergy.class, new FlavorNBTStorage(), FlavoredStorage::new);
-		cm.register(WorldChunkEnergy.class, new NBTSerializableCapabilityStorage<>(), WorldChunkEnergy::new);
-		cm.register(PlayerEnergy.class, new NBTSerializableCapabilityStorage<>(), PlayerEnergy::new);
+		cm.register(WorldChunkEnergy.class, new NBTCapabilityStorage<>(), WorldChunkEnergy::new);
+		cm.register(PlayerEnergy.class, new NBTCapabilityStorage<>(), PlayerEnergy::new);
 		EVENT_BUS.register(MCFluxEvents.INSTANCE);
 		MCFluxResources.preInit();
 		InjectFluxable.registerWrappers();
@@ -72,6 +79,19 @@ public class MCFlux {
 	@Mod.EventHandler
 	public void loadComplete(FMLLoadCompleteEvent e) {
 		EVENT_BUS.register(InjectWrappers.EVENTS);
+	}
+
+	@Mod.EventHandler
+	public void serverStopped(FMLServerStoppedEvent e) {
+		if (MC_DIR == null) {
+			L.warn("Can't save error messages!");
+			return;
+		}
+		try {
+			ErrorReport.saveAllErrors(MC_DIR);
+		} catch (Exception x) {
+			L.warn(x);
+		}
 	}
 
 	private void registerAllInjects(ASMDataTable asdt) {
@@ -96,5 +116,29 @@ public class MCFlux {
 			}
 		}
 		L.info("Registered " + cnt + " inject registries");
+	}
+
+	private static void updateCheck() {
+		ComparableVersion ccv = new ComparableVersion(R.MF_VERSION);
+		try {
+			URL url = new URL("https", "gist.githubusercontent.com", 443, "/Szewek/97a48d6a61b29171938abf2f6bf9f985/raw/versions.json", null);
+			HttpURLConnection huc = (HttpURLConnection) url.openConnection();
+			InputStreamReader isr = new InputStreamReader(huc.getInputStream());
+			JsonObject je = new JsonParser().parse(isr).getAsJsonObject();
+			JsonObject jmc = je.getAsJsonObject("mc");
+			if (jmc == null)
+				return;
+			JsonPrimitive jv = jmc.getAsJsonPrimitive(Loader.MC_VERSION);
+			if (jv == null)
+				return;
+			String v = jv.getAsString();
+			if (new ComparableVersion(v).compareTo(ccv) > 0) {
+				L.info("A newer Minecraft-Flux version is available (" + v + ")");
+				NEWER_VERSION = v;
+			}
+			UPDATE_CHECK_FINISHED = true;
+		} catch (Throwable t) {
+			L.warn(t);
+		}
 	}
 }