+++ src/main/java/szewek/mcflux/config/MCFluxConfig.java
@@ -7,7 +7,7 @@ import java.io.File;
 
 public class MCFluxConfig {
 	public static int CFG_EU_VALUE = 4, CFG_MKJ_VALUE = 100, WORLDCHUNK_CAP = 20000000, ENERGY_DIST_TRANS = 1000000, CHUNK_CHARGER_TRANS = 2000000, FURNACE_CAP = 50000, MOB_SPAWNER_USE = 500;
-	public static boolean ONLINE_ERROR_REPORT = true;
+	public static boolean ONLINE_ERROR_REPORT = true, UPDATE_CHECK = true;
 	private static Configuration config;
 
 	public static void makeConfig(File file) {
@@ -30,6 +30,7 @@ public class MCFluxConfig {
 		FURNACE_CAP = cfgInt("furnaceCapacity", 50000, 1000, Integer.MAX_VALUE, "Energy capacity for Vanilla Furnace");
 		MOB_SPAWNER_USE = cfgInt("mobSpawnerEnergyUse", 500, 100, Integer.MAX_VALUE, "Energy needed for Vanilla Monster Spawner to speed up");
 		ONLINE_ERROR_REPORT = cfgBool("onlineErrorReport", true, "Reports all capability-related crashes made by Minedraft-Flux online");
+		UPDATE_CHECK = cfgBool("updateCheck", true, "Checks if a newer Minecraft-Flux version is available");
 		if (config.hasChanged())
 			config.save();
 	}