+++ src/main/java/szewek/mcflux/network/msg/MsgNewVersion.java
@@ -1,0 +1,37 @@
+package szewek.mcflux.network.msg;
+
+import net.minecraft.client.resources.I18n;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.network.PacketBuffer;
+import net.minecraft.util.text.ITextComponent;
+import net.minecraftforge.fml.relauncher.Side;
+import szewek.mcflux.L;
+
+import java.io.IOException;
+
+public class MsgNewVersion extends FragileMsg {
+	private String version = null;
+
+	public static MsgNewVersion with(String v) {
+		MsgNewVersion mnv = new MsgNewVersion();
+		mnv.version = v;
+		mnv.broken = false;
+		return mnv;
+	}
+
+	@Override public void processMsg(PacketBuffer pb, EntityPlayer p, Side s) throws IOException {
+		if (pb.readableBytes() < 2) {
+			L.warn("Incompatible length");
+		}
+		version = pb.readStringFromBuffer(32767);
+		broken = false;
+		if (s == Side.CLIENT)
+			p.addChatComponentMessage(ITextComponent.Serializer.jsonToComponent(I18n.format("mcflux.update.newversion", version)));
+	}
+
+	@Override public void saveBuffer(PacketBuffer pb) throws IOException {
+		if (broken)
+			return;
+		pb.writeString(version);
+	}
+}