+++ src/main/java/szewek/mcflux/util/error/ErrMsg.java
@@ -1,10 +1,15 @@
 package szewek.mcflux.util.error;
 
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.Set;
+
 public abstract class ErrMsg {
-	private static final long ERRORS_TIME = 5000;
+	private static final long ERRORS_TIME = 15000;
 	protected final String name;
 	protected final Class<?> cl;
-	protected final Throwable msgThrown;
+	protected final Set<Throwable> setOfThrown;
+	public final Throwable msgThrown;
 	protected int cachedHash;
 	private int count, lastCount;
 	private long nextShow;
@@ -12,9 +17,11 @@ public abstract class ErrMsg {
 	protected ErrMsg(String name, Class<?> cl, Throwable thrown) {
 		this.name = name;
 		this.cl = cl;
+		setOfThrown = new HashSet<>();
 		msgThrown = thrown;
 		count = 0;
 		cachedHash = (this.getClass().hashCode() << 24) + (cl.hashCode() << 16) + name.hashCode();
+		addThrowable(thrown);
 	}
 
 	@Override public int hashCode() {
@@ -25,21 +32,33 @@ public abstract class ErrMsg {
 		return obj == this || obj instanceof ErrMsg && obj.hashCode() == cachedHash;
 	}
 
+	public void addThrowable(Throwable th) {
+		if (th != null)
+			setOfThrown.add(th);
+	}
+
+	public Set<Throwable> getThrowables() {
+		return Collections.unmodifiableSet(setOfThrown);
+	}
+
 	public void addUp() {
 		long now = System.currentTimeMillis();
-		if (count == 0) {
-			count = 1;
+		count++;
+		if (count == 1) {
 			printError();
 			nextShow = now + ERRORS_TIME;
 			lastCount = 1;
 			return;
 		}
-		count++;
 		if (nextShow < now) {
 			printShortError(count - lastCount, now - nextShow + ERRORS_TIME);
 			nextShow = now + ERRORS_TIME;
 			lastCount = count;
 		}
+	}
+
+	public String makeInfo() {
+		return "| Name: " + name + "\n| Class: " + cl.getName() + "\n| Count: " + count;
 	}
 
 	protected abstract void printError();